requirements.txt
tgtg
python-dotenv
discord.py